# backend/internal/httpapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "STA Platform API",
    "version": "v1",
    "description": "Auto-generated shallow spec for the STA Go backend. Method, path, tags and security are accurate; request/response schemas are generic except for a few hand-augmented endpoints. Regenerate with `python3 docs/openapi/generate.py`."
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "local"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "sta_session"
      },
      "csrf": {
        "type": "apiKey",
        "in": "header",
        "name": "X-CSRF-Token"
      },
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      },
      "adminMFA": {
        "type": "apiKey",
        "in": "header",
        "name": "X-MFA-Code"
      },
      "serviceToken": {
        "type": "http",
        "scheme": "bearer",
        "description": "extraction / agent service token"
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "error"
        ]
      }
    }
  },
  "paths": {
...
    "/api/v1/admin/experience-revisions/{revisionID}/review": {
      "post": {
        "operationId": "post_api_v1_admin_experience_revisions_revisionID_review",
        "summary": "Create admin experience revisions revisionID review",
        "tags": [
          "admin-experience-revisions"
        ],
        "parameters": [
          {
            "name": "revisionID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": [],
            "csrf": [],
            "adminMFA": []
          },
          {
            "bearer": [],
            "adminMFA": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/experience-revisions/{revisionID}/submit": {
      "post": {
        "operationId": "post_api_v1_experience_revisions_revisionID_submit",
        "summary": "Create experience revisions revisionID submit",
        "tags": [
          "experience-revisions"
        ],
        "parameters": [
          {
            "name": "revisionID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiences": {
      "get": {
        "operationId": "get_api_v1_experiences",
        "summary": "Get experiences",
        "tags": [
          "experiences"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post_api_v1_experiences",
        "summary": "Create experiences",
        "tags": [
          "experiences"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiences/{experienceID}": {
      "get": {
        "operationId": "get_api_v1_experiences_experienceID",
        "summary": "Get experiences experienceID",
        "tags": [
          "experiences"
        ],
        "parameters": [
          {
            "name": "experienceID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiences/{experienceID}/revisions": {
      "post": {
        "operationId": "post_api_v1_experiences_experienceID_revisions",
        "summary": "Create experiences experienceID revisions",
        "tags": [
          "experiences"
        ],
        "parameters": [
          {
            "name": "experienceID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/experiences/{experienceID}/unpublish": {
      "post": {
        "operationId": "post_api_v1_experiences_experienceID_unpublish",
        "summary": "Create experiences experienceID unpublish",
        "tags": [
          "experiences"
        ],
        "parameters": [
          {
            "name": "experienceID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forum/spaces": {
      "get": {
        "operationId": "get_api_v1_forum_spaces",
        "summary": "Get forum spaces",
        "tags": [
          "forum"
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forum/spaces/{spaceID}/join": {
      "post": {
        "operationId": "post_api_v1_forum_spaces_spaceID_join",
        "summary": "Create forum spaces spaceID join",
        "tags": [
          "forum"
        ],
        "parameters": [
          {
            "name": "spaceID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forum/spaces/{spaceID}/leave": {
      "post": {
        "operationId": "post_api_v1_forum_spaces_spaceID_leave",
        "summary": "Create forum spaces spaceID leave",
        "tags": [
          "forum"
        ],
        "parameters": [
          {
            "name": "spaceID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forum/spaces/{spaceID}/threads": {
      "get": {
        "operationId": "get_api_v1_forum_spaces_spaceID_threads",
        "summary": "Get forum spaces spaceID threads",
        "tags": [
          "forum"
        ],
        "parameters": [
          {
            "name": "spaceID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post_api_v1_forum_spaces_spaceID_threads",
        "summary": "Create forum spaces spaceID threads",
        "tags": [
          "forum"
        ],
        "parameters": [
          {
            "name": "spaceID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/forum/threads/{threadID}/posts": {
      "get": {
        "operationId": "get_api_v1_forum_threads_threadID_posts",
        "summary": "Get forum threads threadID posts",
        "tags": [
          "forum"
        ],
        "parameters": [
          {
            "name": "threadID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post_api_v1_forum_threads_threadID_posts",
        "summary": "Create forum threads threadID posts",
        "tags": [
          "forum"
        ],
        "parameters": [
          {
            "name": "threadID",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "security": [
          {
            "sessionCookie": []
          },
          {
            "bearer": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "4XX": {
            "description": "Client error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}